+++ interpreter/sql_wasm.go
@@ -37,3 +37,4 @@ func builtinSQLClose(i *Interpreter, args []Value) (Value, error)       { return
 func builtinSQLMigrate(i *Interpreter, args []Value) (Value, error)     { return Value{}, errSQLUnsupported }
 func builtinSQLTransaction(i *Interpreter, args []Value) (Value, error) { return Value{}, errSQLUnsupported }
 func builtinSQLInsert(i *Interpreter, args []Value) (Value, error)      { return Value{}, errSQLUnsupported }
+func builtinSQLUpsert(i *Interpreter, args []Value) (Value, error)      { return Value{}, errSQLUnsupported }